+++ src/main/java/com/easypost/model/BaseCreditCard.java
@@ -2,6 +2,13 @@ package com.easypost.model;
 
 import com.easypost.net.EasyPostResource;
 
+/**
+ * BaseCreditCard is a model class that represents the base of any credit card.
+ *
+ * @deprecated Use {@link com.easypost.model.PaymentMethodObject} instead.
+ * Last working version: v5.5.0. Removal: v7.0.0.
+ */
+@Deprecated
 public class BaseCreditCard extends EasyPostResource {
     private String id;
     private String object;