+++ src/main/java/com/easypost/model/CreditCard.java
@@ -6,6 +6,13 @@ import com.easypost.exception.EasyPostException;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * CreditCard is a model class that represents a credit card.
+ *
+ * @deprecated Use {@link com.easypost.model.PaymentMethodObject} instead.
+ * Last working version: v5.5.0. Removal: v7.0.0.
+ */
+@Deprecated
 public class CreditCard extends BaseCreditCard {
     /**
      * Fund your EasyPost wallet by charging your primary or secondary card on file.
@@ -14,7 +21,10 @@ public class CreditCard extends BaseCreditCard {
      * @param primaryOrSecondary primary or secondary payment method.
      * @return CreditCardFund object.
      * @throws EasyPostException when the request fails.
+     * @deprecated Use {@link com.easypost.model.Billing#fundWallet(String, PaymentMethod.Priority)} instead.
+     * Last working version: v5.5.0. Removal: v7.0.0.
      */
+    @Deprecated
     public static boolean fund(String amount, CreditCardPriority primaryOrSecondary) throws EasyPostException {
         return fund(amount, primaryOrSecondary, null);
     }
@@ -27,7 +37,10 @@ public class CreditCard extends BaseCreditCard {
      * @param apiKey             API key to use in request (overrides default API key).
      * @return CreditCardFund object.
      * @throws EasyPostException when the request fails.
+     * @deprecated Use {@link com.easypost.model.Billing#fundWallet(String, PaymentMethod.Priority, String)} instead.
+     * Last working version: v5.5.0. Removal: v7.0.0.
      */
+    @Deprecated
     public static boolean fund(String amount, CreditCardPriority primaryOrSecondary, String apiKey)
             throws EasyPostException {
         PaymentMethod paymentMethods = PaymentMethod.all();
@@ -48,7 +61,7 @@ public class CreditCard extends BaseCreditCard {
             throw new EasyPostException("The chosen payment method is not a credit card. Please try again.");
         }
 
-        Map<String, Object> params = new HashMap<String, Object>();
+        Map<String, Object> params = new HashMap<>();
         params.put("amount", amount);
 
         // will attempt to serialize the empty response to a CreditCard object (doesn't matter)
@@ -63,7 +76,10 @@ public class CreditCard extends BaseCreditCard {
      *
      * @param creditCardId ID of credit card to delete.
      * @throws EasyPostException when the request fails.
+     * @deprecated Use {@link com.easypost.model.Billing#deletePaymentMethod(PaymentMethod.Priority)} instead.
+     * Last working version: v5.5.0. Removal: v7.0.0.
      */
+    @Deprecated
     public static void delete(String creditCardId) throws EasyPostException {
         delete(creditCardId, null);
     }
@@ -74,7 +90,10 @@ public class CreditCard extends BaseCreditCard {
      * @param creditCardId ID of credit card to delete.
      * @param apiKey       API key to use in request (overrides default API key).
      * @throws EasyPostException when the request fails.
+     * @deprecated Use {@link com.easypost.model.Billing#deletePaymentMethod(PaymentMethod.Priority, String)} instead.
+     * Last working version: v5.5.0. Removal: v7.0.0.
      */
+    @Deprecated
     public static void delete(String creditCardId, String apiKey) throws EasyPostException {
         request(RequestMethod.DELETE, String.format("%s/%s/%s", EasyPost.API_BASE, "credit_cards", creditCardId), null,
                 CreditCard.class, apiKey);