+++ src/main/java/com/easypost/model/SecondaryPaymentMethod.java
@@ -1,3 +1,10 @@
 package com.easypost.model;
 
+/**
+ * SecondaryPaymentMethod is a model class that represents the secondary payment method.
+ *
+ * @deprecated Use {@link com.easypost.model.PaymentMethodObject} instead.
+ * Last working version: v5.5.0. Removal: v7.0.0.
+ */
+@Deprecated
 public class SecondaryPaymentMethod extends BaseCreditCard {}