+++ src/main/java/com/easypost/model/beta/CreditCard.java
@@ -2,4 +2,11 @@ package com.easypost.model.beta;
 
 import com.easypost.model.BaseCreditCard;
 
+/**
+ * CreditCard is a model class that represents a credit card.
+ *
+ * @deprecated Use {@link com.easypost.model.PaymentMethodObject} instead.
+ * Last working version: v5.5.0. Removal: v7.0.0.
+ */
+@Deprecated
 public class CreditCard extends BaseCreditCard {}